Add support for partial tests
Summary:
There are use cases where operators would like to run one of our offered network test suites on a segment or single node/link in the network. This diff adds the ability to add an optional `whitelist` parameter to schedule or execution creation.

The service expects the whitelist to be a list of strings.

- For parallel/sequential test, the whitelist should be a list of links that we want to test. If no list is provided, it is assumed to be a holistic test
- For multihop test, the whitelist should be a list of nodes that we want to test. If no list is provided, it is assumed to be a holistic test.

In addition, I've created a new dataclass called `TestAsset` to cleanly encapsulate the information. There is an optional field for `link_name` which will help us when we need to query Prometheus for the firmware stats in the future.

Reviewed By: spurav

Differential Revision: D20906434

fbshipit-source-id: b37ae6f1ad0b7c4645eaa03becf594f908ef6bba