Don't permit changing schedule test type
Summary:
The `whitelist` test parameter is tightly coupled to the test type. If the test is parallel or sequential, it should be a list of link names but if it is multihop then the whitelist should be a list of node names. Rather than enforcing that both must be changed if the test type is changed, I am wholesale not supporting the changing of a recurring schedule's test type.

In addition, I've added checks to verify that a network name is valid in the network test API before trying to start a recurring schedule or ad hoc test.

Reviewed By: spurav

Differential Revision: D21006677

fbshipit-source-id: c886726c3322a9eef0c962a67f1e62b54568d4ac