Add dialect option for sequelize orm (#508)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/508

Pull Request resolved: https://github.com/facebookincubator/magma/pull/1521

To more align with Magma technical requirements, we need to support mariadb rather than just mysql.

Sequelize added support for mariadb dialect in v5, and this diff just adds support to choose the dialect as an envioronment variable. Additionally, we now use a mariadb image in the development docker-compose file.

Reviewed By: rckclmbr

Differential Revision: D21077649

fbshipit-source-id: c4bd9260a51c3572ded0b9e961a5b6366bfe6125