Fixing bug in navigation
Summary:
In case of View with no MenuItem (i.e. PermissionsGroupCard) if a View with MenuItem is defined in a later position in the views array, clicking the later MenuItem will not navigate to the correct View.

Plus, removing redundant use of `forEach` and replacing it with `map`&`filter` (according to Flows support group recommendation)

Reviewed By: dlvhdr

Differential Revision: D21118111

fbshipit-source-id: 1a9de3658da22b85caccff126267ad11ee638e7b