Change prod image version to alpine 3.10
Summary:
The `tglib` dev image is fixed to use alpine 3.10 because it is the latest version of alpine that uses python3.7 -- all alpine versions beyond 3.10 use python3.8. We can't use python3.8 for tglib because there is no support for running python3.8 in sandcastle for our tests.

Some tglib-based services (which inherit from the prod image) will need to install their own build dependencies for projects like numpy. In order to do this, developers should be installing the dependencies needed to run in python3.7 which means that they have three options:

1. Inherit from the `tglib` dev image
2. Ask developers to add alpine3.10 to their repositories file and use version pinning techniques.
3. Fix the prod image to alpine3.10

I opted for option 3 because its the most fool-proof solution.

Reviewed By: spurav

Differential Revision: D21135302

fbshipit-source-id: 0b9a7f647f047a198902b8e76adac913dd6e6bc1