cpp backends: pin rsocket-cpp to working commit
Summary:
We use rsocket-cpp master in the build process - because the last release is from 2017.

A commit was pushed yesterday that broke compatibility with folly.
https://github.com/rsocket/rsocket-cpp/commit/a520ddfba9f817e0cb179df8b59bdd55809b1b2b
https://github.com/facebook/folly/commit/6e85117e8c9adb6a0e25d8c1226ebfc3886a97ab

We would need to use latest (not a released version) of folly with rsocket-cpp master to build, or pin rsocket-cpp so we don't have any unexpected breakages.

Reviewed By: omikader

Differential Revision: D21283873

fbshipit-source-id: 2e3f1d3e4298aaeeaec02075d7c2638536c512a8