misc(ansible): move default gluster_exporter command to Dockerfile
Summary:
According to [literature online](https://phoenixnap.com/kb/docker-cmd-vs-entrypoint), the `ENTRYPOINT` keyword is preferred when you want to define a container with a specific executable.

As this service is a single executable copied into the `gluster` image, it makes sense to use `ENTRYPOINT` as opposed to supplying a command in the compose file.

Reviewed By: pmccut

Differential Revision: D21394873

fbshipit-source-id: 5419eca86797eb0f0b31bc2f6c9eba021fe2d7cb