Remove snackbar dependency in fbcnms-alarms
Summary: Snackbar dependency was a little more involved, so I split it off to its own diff. There was a dependency on the Design System for `Text` which is just a wrapper around `<span>` so I just replaced `Text` elements with `<span>` and the props that `Text` would have passed in to them.

Reviewed By: aclave1

Differential Revision: D21355580

fbshipit-source-id: 58d04d931fba49d3c82b1ca71ab7372d8b3ea4ff