feat(network_test): add schedule/execution filtering for partial vs holistic tests
Summary:
cjminer requested that schedules/executions be additionally filterable by whether or not they are run partially or holistically on the network. This required adding `WHERE` clauses on the `whitelist` column.

This was challenging because JSON columns will use `null` instead of the MySQL `NULL`  so I added `none_as_null` to the schema. This allows me to use the `WHERE` clauses correctly.

Reviewed By: spurav

Differential Revision: D21406243

fbshipit-source-id: 3ce8290d63b59fb414cc414153fd92c0f9ba9476