misc(network_test): generalize link_name column to asset_name
Summary:
We save the `link_name` in order to fetch the proper firmware stats during parallel and sequential link tests. The `link_name` is also used by the NMS for displaying the results in the map overlays.

In order to display the results for multihop tests in the map overlays, I've generalized the column to be called `asset_name` and supplied the name of the node under test as the value.

Differential Revision: D21412158

fbshipit-source-id: 5110f40a81a1c78e9b4696887197ea5cbd53b26e