Gauge CN power status
Summary:
This service, part of "analytics", determines with CNs have been powered off.  The algorithm is described in the README.md for analytics (which is not part of this diff).  At the highest level, the algorithm collects 30 minutes of stats from Prometheus and checks all CNs in the network.  If they are up then certainly they are not powered off.  If the DN is reachable and the E2E is trying to ignite the link and getting no response over the 30 minutes, then the CN is considered powered off.

possible states are:
    INIT_UNREACHABLE = 1  // means that the initiator is not reachable
    NO_ASSOC_REQS = 2  // means that the E2E is not trying to bring the link up
    LINK_ALIVE = 3  // means that the link is alive (so not powered off)
    CN_POWERED_OFF = 4 // means that the CN is powered off
    UNKNOWN = 5   // this should never happen - it would indicate a bug in the algorithm

The results are written to Prometheus as a 30s stat called "cn_power_status"

I am thinking that the pipeline should run every 5 minutes. (the test below runs every 30s so I would get results more quickly).

Reviewed By: omikader

Differential Revision: D18065128

fbshipit-source-id: ad3c868ce9d23a37f7750e14762bbda5ecb9da2d