misc(tglib): drop direct dependencies on non async client libs
Summary:
The latest version of `aiokafka` requires a version of `kafka-python>=2.0.0` which was causing the tglib dependencies to install incorrectly.

The original reason I was including `kafka-python` and `pymysql` as dependencies was because their async library counterparts use the exceptions defined in the non async libraries but I realize now that I can import these errors from the async libraries directly.

Reviewed By: spurav

Differential Revision: D21595890

fbshipit-source-id: 4b7f9ff8eccc338cf901e0a106d4a3c6d37a8436