drop statistic timestamps in the future, batch kafka consumer poll
Summary:
Some PUMA setups are seeing node timestamps that are June, which was causing us to write link health/availability in the future, then getting confused when their timestamp was later fixed.

This drops and logs the # of messages that were invalid for that poll

Also, batch polling because it should be faster.

Reviewed By: omikader

Differential Revision: D21649041

fbshipit-source-id: 08a3a5a3271ae177e9f6d7f64ba04b72b6897d6a