Fix facebookexternal magma flow (#645)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/645

Pull Request resolved: https://github.com/facebookexternal/terragraph-nms/pull/275

Pull Request resolved: https://github.com/facebookincubator/magma/pull/1704

Pull Request resolved: https://github.com/facebookexternal/fbc/pull/2482

The installed dependencies in external all live in the root directory. We need to add relay as a "declaration" - i.e. use the flow types, but don't check that they compile.

Also add `prepend-file` to the relay compiler because it's used and needed but not listed, resulting in build errors. (It was probably fine in our directory since it's a dependency of another package).

Reviewed By: rckclmbr

Differential Revision: D21670505

fbshipit-source-id: a61ddc22328a1c0310c69a531177c475d3deb99d