Save scan data to database
Summary: While testing D21627656, I had commented `write_scan_data` function call because my database was on a different version (cause of Scheduler diffs), and also because saving results to the db was not the focus of that diff. Was supposed to uncomment before shipping, which I missed. Fixing it in this diff.

Reviewed By: hmehmood0

Differential Revision: D21680396

fbshipit-source-id: 520556706cff2eac83b7b2960caf3c564468d8bc