Use provisioning instead of grafana_utils
Summary:
See D21463719, this updates the dev docker-compose to do the same thing.

This requires duplicating all the provisioning files which isn't great, is there a plan for removing this setup? It seems like everything can/should be done through the nms installer since that's what is actually used in the end

Reviewed By: omikader

Differential Revision: D21652332

fbshipit-source-id: 8ed65d6155f1584fd984fb580e2027678ec15416