feat(nms): Adding select key and functionality to schedule table
Summary: Since the table has IDs of executions and schedules, we couldn't use the ID column for selecting the table because sometimes it would double select. Also couldn't use numbers because of the way custom table is made, it takes all row values and then checks if any of them are euqual to select key value, so it would again double select. That is why it is a string with 'table' + index in the array.

Reviewed By: aclave1

Differential Revision: D21650065

fbshipit-source-id: c6acb7b39acf2894f5540c711d3ad4c0ea810d3d