Analyze node alignment
Summary:
Analyze node alignment
This job analyzes alignment for every link in both directions (A and Z) for a network. Tx and Rx beam index is retrieved for both ends of the link. The beam indices are converted to degrees and compared with the misalignment threshold. The difference between tx and rx degree is calculated and compared with the tx and rx degree difference threshold value.

Reviewed By: hmehmood0

Differential Revision: D21695468

fbshipit-source-id: b64a9eff626f781f6a3fb99dd774184aa184e3fe