misc(tglib): add numpy to testing environment
Summary:
Some `tglib` services had `ptr` disabled by default because they use `numpy` and it was difficult to install `numpy` to our testing environment. This diff installs `numpy` to the virtual environment and enables `ptr` for those services.

In order for `numpy` to work, the directory to which the C-extensions are installed must be added to the `$LD_LIBRARY_PATH` on the system. Since sourcing shell scripts doesn't preserve environment variables between steps on sandcastle, I manually exported the variable as part of the `ptr` run step.

The `setup_tests.sh` script will now need to be sourced on the devserver in order to properly export the `$LD_LIBRARY_PATH`.

Reviewed By: cooperlees

Differential Revision: D21952606

fbshipit-source-id: 1a2e7f1f103c27b8952b0ef27f18ddb74879c229