misc(udp_pinger): rename files for consistency
Summary:
I want to add information about the `udp_pinger` to our operator docs. Before I do that, I think we should resolve all of the naming inconsistencies. pmccut and I agreed that the best name going forward should be the `udp_pinger`.

Changes:
- All Prometheus metrics generated by the service will be prefixed with `udp_pinger_` instead of `pinger_`
- The path created on the host machine is changed from `.../udp_ping_client` to `.../udp_pinger`
- The systemd service responsible for restarting the service if it goes down is renamed from `udp_ping_client_monitor` to `udp_pinger_monitor`
- Both the container itself and ansible role are renamed from `udp_ping_client` to `udp_pinger`
- The source code is moved from `ping_service` to `udp_pinger`

I've kept the binary name as `udp_ping_client` in order to be consistent with `UdpPingClient.cpp` where the `main` function is.

Reviewed By: pmccut

Differential Revision: D21860656

fbshipit-source-id: 53511a4fc63cef5b6b4938d7e61d0c7cb2a775e1