Move activity log to async service and control move with feature flag (#749)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/749

First, open `work_order_activities_hook` completely. It was already opened for partners with no issues.
Second, create a new feature flag `move_work_order_activities_to_async_service` to determine if it the events should be handled inside the hook as it is now or inside the async service.
Also remove `graph_event_logging` and just use `move_work_order_activities_to_async_service`. If it's on just send the event and the async service will just handle it if it's a work order event.

Temporal changes for "Experiement" phase:
* I moved `graph/event/workorder_activities.go` to `graph/event/workorder_activities.go` because I need it both from `graph` and from `async`. Once I open the feature flag, I'll move it to `async` only
* I left `graph/event/workorder_activities_test.go` and copied it to `async/handler/workorder_activities_test.go`. Once I open the feature flag I'll delete `graph/event/workorder_activities_test.go`

Reviewed By: a8m

Differential Revision: D22061264

fbshipit-source-id: 34fd3c3327dae3b56f50c81e09f315d953623465