fix(ansible): do not overwrite db files during reinstallation
Summary:
Some `msa` services have their own databases in MySQL. In order to avoid polluting the the `group_vars` with default credentials for every service, there is a step in the `msa` installation role for generating a random password using ansible's [`lookup`](https://docs.ansible.com/ansible/latest/plugins/lookup/password.html) plugin when templating `msa_db.env`.

The plugin saves the file to some temporary location on the installer machine in `/tmp` in case it needs to be re-sourced, which happens later when templating the `alembic.ini` file.

If the installer is run again from the same installation host, and the file still exists in `/tmp`, then the same password will be reused when templating the two files which works fine, however if the installer is run again from a different host, then

- the files will get re-templated with a new password because `/tmp` is not the same
- the files will be overwritten on the target machines because we always `force: yes` those files.

The problem with this is that the actual user and password don't get updated in MySQL because the `create_msa_database.sh` script does not recreate the user. When this happens, the environment files and MySQL become out of sync and we get access issues to the DB.

Reviewed By: aclave1

Differential Revision: D22096463

fbshipit-source-id: ac9564271c28fcb3545f237f4c361fd0b38af1bb