Setup Kubernetes cluster from ansible
Summary:
This adds `kubernetes-install` and `kubernetes-uninstall` to the nms cli. To both commands you can specify a number of master and worker nodes, and it will install docker/kubeadm/kubectl and create a cluster.

It basically does the steps outlined [in the docs](https://kubernetes.io/docs/setup/production-environment/tools/kubeadm/create-cluster-kubeadm) and a bunch of it is copied from [this guide](https://www.digitalocean.com/community/tutorials/how-to-create-a-kubernetes-cluster-using-kubeadm-on-ubuntu-18-04).

Reviewed By: aclave1

Differential Revision: D22176760

fbshipit-source-id: 108665ba9c0c7d7c44d0b19c1df7ac0dbcd7088c