Correctly update end time of scan execution
Summary:
End time of any scan execution was being set on every update to `ScanTestExecution` table. Thus, the execution `end_dt` was set when we updated the execution status from QUEUED to RUNNING, which is not accurate behaviour.

Updating the logic to explicitly set the end time of the test when the test has either FINISHED or FAILED.

Reviewed By: hmehmood0

Differential Revision: D22333215

fbshipit-source-id: b7841876a0d701e5648648a6c535a9c717d9eccd