fix(network_test): handle multihop test when if no default routes
Summary:
The `prepare` method for `network_test` prepares the network test assets to be tested from the network topology. Multihop tests are run from nodes to a random PoP in their default routes, however there is no handling for when there are 0 default routes available.

This diff skips any nodes that do not have default routes and logs an error indicating that the intended node was unable to be properly tested.

Reviewed By: spurav

Differential Revision: D22434613

fbshipit-source-id: 63705fc88beb9dfa1e11af1569b960e94146ac58