Separate Kubernetes Ansible setup from nms_stack
Summary:
This is mostly a code move along with some small changes to the nms installer so it can work with the new file names. This will help us keep everything separate from the docker swarm setup so it's easier to differentiate between the two.

It's named `k8s_nms` for now, hopefully later we can delete the swarm installer and rename it to just `nms` or something.

`hg` didn't recognize the code moves, so files that were moved and changed have comments

Reviewed By: aclave1

Differential Revision: D22294738

fbshipit-source-id: f4f8f21b07f1b86916e567e2a47f18c7cd8cb66c