Subscriber's Dashboard & Subscriber Overview/Config Screens (#1952)
Summary:
# Overview
This PR covers the subscribers v2 dashboard, and specific subscriber view Overview/Config screens.

## Subscriber's Dashboard
Not much work needed to be done here, but the designs were flattened and the overall top bars were streamlined to match other sections on the site.

![image](https://user-images.githubusercontent.com/8878152/87437822-c6ea7880-c5bc-11ea-8035-9ed3d9b0bb8b.png)

## Subscriber Overview
Updated the overall layout of the page, including updating the KPI's to make use of the KPI Grid component we now have.

![image](https://user-images.githubusercontent.com/8878152/87438979-344ad900-c5be-11ea-9ec2-73c93c9abffc.png)

## Subscriber Config
Updated the overall layout of the page and improved the way in which collapsable lists look and operate.

![Jul-14-2020 10-42-51](https://user-images.githubusercontent.com/8878152/87443095-0c11a900-c5c3-11ea-8076-79e2d70e1e59.gif)

## Odds and Ends
In addition to the overall core changes mentioned above, there have also been some minor style bug fixes throughout the files.

Pull Request resolved: https://github.com/facebookincubator/magma/pull/1952

Reviewed By: xjtian

Differential Revision: D22537578

Pulled By: karthiksubraveti

fbshipit-source-id: c0bb97d05a30ec5aa2faca60f6dfe911e7995ae6