Restrict accessible tabs per project
Summary:
A bug was recently encountered in `magmalte` where if 'Inventory' was selected as an accessible tab, when visiting `https://<org-name>.localhost` the page would fail to load because there is no router entry for inventory paths in magmalte.

Thus, it makes no sense for inventory to be selectable as an option. This applies for all the other tabs that aren't "NMS" (automation, hub, etc.). Until those routers are implemented in magmalte they are useless.

Since projects should be independent of each other, the available tabs should be different for each. This diff allows a prop to be passed to the `OrganizationEdit` so you can specify specific tabs that are accessible. This will have no impact on symphony's deployment, but it restricts `magmalte` to only have "NMS" as a tab option.

Reviewed By: naor9991

Differential Revision: D22467079

fbshipit-source-id: e82be6d5472425eb1af47642ff3402bf9d6f0755