Add build_zip command to generate binary of all resources
Summary: This adds a `build_zip` command to `setup.py` to create a binary of all the files needed to configure the cluster. It runs `jinja2` to template out files as needed, then adds them all to a single zip file. Since everything is a Kubernetes resource, they can all just be `apply`ed, and Kubernetes will figure out which parts to change.

Reviewed By: aclave1

Differential Revision: D22522189

fbshipit-source-id: 1843acc923b5e03d264ee33056f9a0c483b04ec7