fix(network_test): update execution status to failed if still processing
Summary:
Stopping a network test execution can happen in one of two ways. The execution is manually aborted by the user, or the execution timeout logic has kicked in.

In the case of manual abort:
- All of the individual sessions should be marked as `ABORTED` in the database
- The execution itself should be marked as `ABORTED` in the database

In the case of timeout logic:
- All of the individual sessions should be marked as `PROCESSING` in the database
- The execution itself should be marked as `PROCESSING`
- Then, we sleep for the timeout duration
- Upon waking up, any individual sessions still `PROCESSING` should be marked as `FAILED` and the execution itself should be marked as `FAILED` if all of the sessions `FAILED`, otherwise `FINISHED`.

Before this diff, individual sessions in the second case were never properly being updated to `FAILED` if they were still `PROCESSING` because it was looking for sessions that were in the `RUNNING` state.

Reviewed By: spurav

Differential Revision: D22514260

fbshipit-source-id: 4fcee7a7b268061fb4ea08a8d756b3b1e3d79f8d