Asyncify grafana handler to improve performance (#16)
Summary:
Pull Request resolved: https://github.com/facebookincubator/fbc-js-core/pull/16

Loading the grafana page can take a few seconds which feels laggy. This diff takes the function `syncTenants` and makes it run async with the other syncing functions since the result of this is independent of all the others. It also happens to be the longest and in testing it cut the load time in about half.

Reviewed By: rckclmbr

Differential Revision: D22736306

fbshipit-source-id: 083c8a745b5be10a482019641d085293379cc1b3