fix(nms): resolve race condition when editing node name and MAC addresses
Summary:
Editing all of the node fields had nondeterministic success due to the usage of `Promise.all` to issue various e2e requests to the controller. There are three different types of changes to a node:

1. Changing the node MAC
2. Changing the node's radio MACs
3. Changing anything else about the node (node name, is_primary, etc.)

Change types 1 and 2 require the name of the node involved in the change. If the node name is also changed in type 3, then the order in which the requests are processed becomes extremely important.

Either changes 1 & 2 must happen first (in any order) with the original node name, or change 3 must happen first followed by changes 1 & 2 (in any order) with the new node name.

This diff changes the the `preConfirm` anonymous function to always run changes 1/2 before change 3.

Reviewed By: aclave1

Differential Revision: D22737792

fbshipit-source-id: ce4d7042cf087fe5be7427558ce9bc373b9ef774