Add resources for database
Summary:
As a test run of adding a service to the cluster this adds the MySQL database with its init scripts and such.

For storage this config specifies a 10 Gi PersistentVolume backed by local disk, but it should eventually be using a StatefulSet + MySQLs replication.

Taken mostly from https://kubernetes.io/docs/tasks/run-application/run-single-instance-stateful-application/

Reviewed By: aclave1

Differential Revision: D22474372

fbshipit-source-id: 8436273da92790ae3238b3c3e3b2752cee430e2d