Optimize topology fetching for analysis
Summary:
We need link_name to node_mac mapping and node_mac to node_name mapping to perform scan analysis. So far, the service was fetching latest topology using the api service for every token response we get over Kafka, which is not optimal.

Now, we fetch the latest topology once, at the start of every scan. This way, each scan has the latest topology info.

Reviewed By: hmehmood0

Differential Revision: D22877729

fbshipit-source-id: e160071cb9f5bf69a4dcb4469b45cbe788c281eb