Improve CWF queries and labelling (#27)
Summary:
Pull Request resolved: https://github.com/facebookincubator/fbc-js-core/pull/27

Some of the dashboards were not behaving as expected. Most of this was queries that performed aggregations not showing the relevant labels. This was because multiple aggregation results were returned which caused the label to not be shown. This diff fixes those and some other nits. Also added units to most dashboards.

Reviewed By: mpgermano

Differential Revision: D22985171

fbshipit-source-id: 24b0e46ebdd9195f4b150d98c321d9f345d9838e