Improve response for when we fail to start scan
Summary:
When we start a scan test using the service, we get the following response:
```
Started new scan test execution with ID: {execution_id}
```
We get this response even if there was any error when starting the scan.
Now, we will get the explicit error response if the service was unable to start the requested scan.

Reviewed By: omikader

Differential Revision: D23014663

fbshipit-source-id: 9fb72a1208070544090e1cd37a10e929cccc1979