fix e2e start scripts from py -> lua cli switch
Summary:
The start scripts use .py, but these were just deleted in D23043126 (which means existing start scripts break with M57)

The lua scripts were added in M47, so this makes M47 the oldest version that will work with this new script.

If that's a concern for folks, then we would need to make a check in the script to support both.

Reviewed By: omikader

Differential Revision: D23099004

fbshipit-source-id: 5e8818b7c89809fe4b60bc9704c0dddca9070749