BE(tgnms): Flow-typed UI Bootstrapping
Summary:
When the NMS UI HTML page is rendered, information such as env-vars, the user, etc is serialized as json. The NMS UI then reads this data at startup.

This extracts that logic to a separate file with a unit test and flow-types it. The UI then uses these same flow-types.

Reviewed By: cjminer505

Differential Revision: D23145062

fbshipit-source-id: 08e93a8be309bc02abe03fd90f00290510a3a7d3