fix(tgnms): disable all websockets code
Summary: Currently, I believe the websockets library we're using is the cause of the recent NMS crashes. This diff disables usage of the library so we can test if that's the cause of the issue. The websockets code needs to be improved anway so soon it'll be a good time to replace the library.

Reviewed By: cjminer505

Differential Revision: D23144068

fbshipit-source-id: ac5ce4756a1b0722c7c963d5701ee82bd4ce93f1