feat(nms):BE:when you change to another tab, switch to default map mode
Summary: There was a problem when in test or scan mode and you switched to another page, scan and test ID would be removed from the URL but the mapMode context would still see the mapMode set to test or scan mode. Now, if scan or test panel load and dont see the mapmode in the url, it sets mapmode to default.

Reviewed By: aclave1

Differential Revision: D23218256

fbshipit-source-id: 06b0412483f1e43fa81c19827f0a5072170a5fc9