add aiomsa library
Summary:
This copies `tglib` to an `aiomsa` library that is intended to serve as a base for both Maveric and Terragraph projects. Most of the functionality is preserved directly, some of the Terragraph specific things such as APIServiceClient and KafkaProducer (which was tightly tied to Terragraph thrift definitions) are removed. The way clients are registered is also changed so that any clients can be added instead of just the ones `tglib` knows about.

This is what we are going to open source. For now this exists alongside `tglib` (so there are 2 copies of the code), tglib will retain its specific clients but have the duplicated parts removed once we get a review from the rest of the team (D23269555).

Reviewed By: al8

Differential Revision: D23269557

fbshipit-source-id: fb6efdf370f3eba4e74f70e72854fe2105e36fe2