fix(tgnms): Fix NMS settings crash
Summary:
fixes the crash

The real issue isn't firefox vs chrome. Paul and I both use firefox as our personal browsers. Firefox was autofilling the "nodeupdate auth token" field in the settings form since it's the first "password" input type on the webpage. The real issue is that the secret fields have been broken for a while and I never caught it because i didn't have my env configured the same way as paul's.

Reviewed By: pmccut

Differential Revision: D23374304

fbshipit-source-id: 1303bfb1ea276f3315b26926cc051f419d90ed6a