Fix URL matching regexes
Summary:
The regexes we were using to route to separate services did not match URLs
without a trailing slash (e.g. hostname/grafana/ would work, but
hostname/grafana would redirect to NMS)

This also extracts some stuff into variables so it can be changed in k8s (the
DNS resolver is different, and the upstream DNS names are also
Kubernetes specific). This lets us keep it all as one file that works in both
stacks.

Reviewed By: aclave1

Differential Revision: D23273487

fbshipit-source-id: 9eb20adbc4e8b3a6f045119874268f912107e917