feat(tgnms): Map Profiles, MCS Estimate editor
Summary:
Creates a "Map Profile" editor which allows for the MCS estimate to be configurable.

The default profile cannot be edited, but it can be copied to create a new profile.
https://pxl.cl/1k1bD

Multiple networks can be assigned to a profile
https://pxl.cl/1k1bz

Here's the estimates for the "Louisiana" map profile, the rings are much smaller
https://pxl.cl/1k1bk

Here's the estimates for the default profile
https://pxl.cl/1k1bn

A basic Mcs estimate editor
https://pxl.cl/1k1bq

https://pxl.cl/1k1bv

Reviewed By: cjminer505

Differential Revision: D23442674

fbshipit-source-id: d0b49966acbb0db2bb147d5a2f63cacf2cc5e05a