Network health service
Summary:
This is the first diff in a series of diffs that introduces `network_health_service` to our suite of services.
In this version, we have a producer-consumer `tglib` service that has a job that fetches stats from `prometheus`, `network_test`, `query_service` ` and `scan_service`.

Please refer [this quip](https://fb.quip.com/bOMCAYUaaDpB) for all stats used by this service

Reviewed By: hmehmood0

Differential Revision: D22158690

fbshipit-source-id: 00e1c591919670c41b3fe7f2adcfac8a52047cba