feat(tgnms): Annotations - set maxZoom on fitBounds call
Summary:
When switching between layers, we fly the user over to the bounds of the annotation layer. If the layer contains a single point annotation, or a smallish polygon, the camera zooms too much to fit the bounding box to the screen. This sets the max zoom to 15(when switching layers).

before
https://pxl.cl/1lb91

after
https://pxl.cl/1lb9B

Reviewed By: mahyali

Differential Revision: D23734278

fbshipit-source-id: b50ceb3f0e683989d4134c1d5bfc78c44378049a