BE(tgnms): Refactor mapbox-draw-gl usage, update to new apis
Summary:
Refactor:
DrawLayer was in a weird place. It started out as an isolated component which handled all ajax and mapbox-draw-gl state. As more things needed to be shared by the context, more was extracted from this file. I've decided that this file should only handle toggling draw on/off and the mapbox/maplayer specific parts of that. The rest of the mapbox draw state is handled in useDrawState

Flow updates:
mapbox-gl-draw uses mapbox's builtin event system to notify of changes, this is the only way we can detect what the user does. This builtin event system is actually internal to mapbox and this type of usage is deprecated, so the flow-types don't match up. I extracted as much as possible away from mapbox's events system and the parts that had to stay are a bit cleaner now.

API Updates:
* removes the updateFeatures function on the context with deleteFeature saveFeature

Reviewed By: cjminer505

Differential Revision: D23661042

fbshipit-source-id: 046cfaf844d2ddab59e88b70892205b835ee56b0