Add rage command
Summary:
This adds the rage tracking from the NMS CLI to the Kubernetes installer. It also edits the rage tracking to include the config used to run, which will be helpful for debugging.

This diff has some more minor fixes and changes to the help text to make it more informative.

Reviewed By: aclave1

Differential Revision: D24032220

fbshipit-source-id: d17d9d726608533c4e66d2428d4d29a5ea5cf2f3