Bundle manifests into installer
Summary: This moves around some folders (again) so that we can include the Kubernetes manifests in the installer Python package. This makes development easier since it defaults to the manifests installed, and initial installation is simpler since you don't have to point the installer to some existing source (the `--template-source` flag is still there, this just adds a set of manifests as the default, so no updates to the installer are necessary to use a different set of manifests).

Reviewed By: aclave1

Differential Revision: D24145838

fbshipit-source-id: 618b8f3acfbab842ea8d4ce621e325c6a9693690