fix(nms): use FULL_UPGRADE type to resolve full upgrade race condition
Summary:
There was an issue in the NMS where full upgrades would occasionally try to run the COMMIT phase before the PREPARE phase due to the use of `Promise.all`.

By using the `FULL_UPGRADE` option, we can avoid this issue since the controller already takes care of the scheduling logic for us.

Reviewed By: aclave1

Differential Revision: D24227014

fbshipit-source-id: c4dbb1a5bea3a7ce64091e167d19ca49778728af