feat(tgnms): NMS Settings defaults
Summary:
* Fixes SettingsEngine to use the `defaultValue` field of a setting def.
* Adds a "reset to X" button to the settings page
* removes the "required" validation and doesn't trigger URL validation on empty fields

Reset button

The NMS's environment variables come from (in order of importance):

* settings.json file
* The environment
* .env file

The presence of an environment variable means that it is defined (LOGIN_ENABLED="" is true). Editing a setting in the Settings UI (and saving to settings.json) defines the setting in the settings.json file and it's merged into the environment. If the user simply backspaces the setting, it will be an empty string but not removed from the settings file, this makes it impossible to ever get back to the fallback value. The reset button deletes a setting definition from the settings.json file.

https://pxl.cl/1nFl8

Reviewed By: cjminer505

Differential Revision: D24427388

fbshipit-source-id: 35559ea7f3e98503e62f1ce5d30ac33e2e3cc46d