fix(network_test): use linkLocal option to ensure desired link is tested in multisector
Summary:
For multisector nodes, the MAC address information in the links are the WLAN MACs, not the node MACs. Network test needs the node MACs in order to start Iperf traffic on the nodes using the E2E API.

This diff changes `link.py` to use the link `a/z_node_name` to get the `node_mac` instead of assuming the value of `a/z_node_mac` are node MACs.

In addition, in order to ensure that link tests on multisector nodes are over the desired link, I've introduced the usage of the optional `useLinkLocal` API param (for link tests only!).

Reviewed By: elludraon

Differential Revision: D24514885

fbshipit-source-id: d4c02ae0fc1e6a5ea4736350c0200b78b01ed23e