Add a Python script to build MSA images
Summary: It's a pain to update all the Jenkins build scripts and keep them all in sync. Some of them have different needs but all the MSA ones are pretty much the same. This adds a script to handle the builds, so the Jenkins build just runs this script and nothing else.

Reviewed By: omikader

Differential Revision: D24523797

fbshipit-source-id: c2cf0d41706f18683e047e29986155aa3a2f2e98