feat(nms): add jenkins script to load SSH key for custom build
Summary:
The `LTS_patch_tagging` Jenkins job will be responsible for tagging each additional commit(s) on the `terragraph-apps` repository and pushing the tags back upstream.

Since this is a custom build, we must load the GitHub SSH key details manually.
{F341789293}

Jenkins has `git version 1.8.3.1` which means I need to use the `GIT_SSH` environment variable to load the key details. Unfortunately, `GIT_SSH` must point to a shell script, and not the actually to a command that loads the SSH key details. To get around that, I found this [solution](https://stackoverflow.com/a/48050119) wherein we set `GIT_SSH_COMMAND` in the shell and the script sources it accordingly.

If we ever upgrade to a `git` version that supports `GIT_SSH_COMMAND`, this solution will still work because `GIT_SSH_COMMAND` will take precedence over `GIT_SSH` if both are set.

Reviewed By: pmccut

Differential Revision: D24553075

fbshipit-source-id: d4235fe69a12d05204900bc79c026888d6f1ae4f