Fix Gluster on ipv4
Summary:
We weren't threading through the variables for `is_ipv6` to the Ansible playbooks themselves, adding them to the `group_vars/all` will cause the templating to include them.

This also makes us backup the configs in a directory since there were already 20+ on the lab vm that were cluttering up the `/opt/terragraph` directory.

Reviewed By: aclave1

Differential Revision: D24555963

fbshipit-source-id: 0b443e8a1600a5021f3f99dd0dfddb0ad3411a36