Sort managers/workers before naming in installer
Summary:
This sorts `managers` and `workers` by name before generating the Ansible inventory (and assigning names like `worker-0`). This way the machine name is fixed between installs for the same cluster.

This also fixes a bug where we were determining whether to use `ipv4` vs `ipv6` based on the installer machine's IP, not the manager node's IP.

Ideally we could sort on the IP, but Ansible doesn't look like it gives an easy way to fetch these for all nodes, so it'd be a lot more complicated than this.

Reviewed By: aclave1

Differential Revision: D24658329

fbshipit-source-id: b588a532cbd30d446cdf0e97db5e0cf542a3b6f1