feat(tgnms): Remote Overlays V0 - Links Layer
Summary:
Allows the user to create "Custom Overlays" for a network profile to request data from an external service. The NMS will query the apis defined by custom overlays and display the data as overlays on the map.  This is a step toward the goal of giving the Wireless Systems team a space to iterate on their custom visualizations without being constrained by the UI team. We are also developing a process for custom visualizations to "graduate" into an overlay distributed by default with the NMS.

https://pxl.cl/1q091
https://pxl.cl/1q0cL
https://pxl.cl/1q094

What does work:
* Rendering link colors and custom text
* Legend UI for links

What needs to be done:

* Rendering site overlays: SitesLayer doesn't support it yet. That will be done in a stacked diff.
* Rendering node overlays, there are currently no "node overlays" in the NMS ui, but this will support that once we have those.
* Displaying extra metadata in the tabular format discussed with shardul/haleema
* Developing the color theming feature so that overlays can have a consistent look and feel despite being created by different developers.

https://pxl.cl/1q0wV

Reviewed By: cjminer505

Differential Revision: D24581286

fbshipit-source-id: 7b0df592b67568b02748989b09e47adc014ef918