fix(query_service): add new radioMac label for multisector
Summary: Multisector nodes are using the WLAN MAC address for the `nodeMac` label in friendly firmware metrics in query service. This diff changes the code such that the node's MAC address is always used for `nodeMac` and the value for  `AggrStats.entity` is assigned to a new label called `radioMac`.

Reviewed By: hmehmood0

Differential Revision: D24740698

fbshipit-source-id: 9b11c9ed28604d613dc7e6936089ead1b944ec28