misc(tgnms): move .env file creation from Dockerfile to CI
Summary: The NMS UI uses `.env` to populate information in the changelog. In moving to the new `registry.py` script for generalizing our build process, I briefly broke this process. This diff moves the commit hash/date logic out of the Dockerfile in favor of doing it directly in Jenkins.

Reviewed By: driazati

Differential Revision: D24892290

fbshipit-source-id: 58c1a7e6d299e78553e271a347ce63d02f67480c