feat(nms): add docs build
Summary:
This adds a setup so we can write docs about NMS in `connectivity-lab` that can be broken up into logical pieces. It includes some sample content just for demo purposes (e.g. how to make sub pages, include an image, link to other docs)

It uses a similar pandoc build to the one in meta-terragraph, which has some rough edges (the `%` needs to be used to specify the page title, we still have to make some manual edits to the generated html).

Per Jeffrey, some general concerns:
* We probably don't need to put these in the operator docs if they are specific usage instructions, PDF size is also a concern with lots of screenshots (< 10 MB is nice for emailing)
* Can we host meta-terragraph docs on NMS? They likely have different NDA, we'll probably have to consult legal or something here.
    * To work around this for now I'm not exposing anything that's not part of nms, so it's just the E2E API, tglib, and NMS docs for now.

Reviewed By: aclave1

Differential Revision: D24774988

fbshipit-source-id: 35a5a4ebd19cc2bd64a1b607b9604af28a25182c