fix(nms): fixed congestion test problem
Summary:
There was a problem where the array was a list of promises rather than strings and if the list was empty, it would still run a test on all the nodes which is not what we want.

When the details panels become functional, if there are no CNs to be found, it would be good to alert stating the test was not run because there were no CN's found that this link serves.

Reviewed By: aclave1

Differential Revision: D25219161

fbshipit-source-id: 771ca26e646ec2ce4427f035cbcf3b34f499dcb4