misc(jupyter): use public jupyter image
Summary:
The `jupyter` image used to be able to load + run `tglib` code, however building this image was such a pain because `tglib` needs lots of dependencies (including `fbthrift`) and we also needed to manually build `numpy`, `scipy`, and `matplotlib`. To my knowledge, no one was really taking advantage of this feature of the service anyways.

1. Use `jupyter/scipy-notebook:latest` image
2. Move `jupyter` from `msa` stack to `nms`
3. Delete code/artifacts referring to custom `jupyter` image

Reviewed By: driazati

Differential Revision: D25445341 (https://github.com/facebookexternal/terragraph-apps/commit/ca6c2249dfccb895916fd5bc23ce17001830b3ef)

fbshipit-source-id: 9c9fec3b5b4fa1a77e62dbf1773912366ad884a0