fix(tgnms): Fix MCS estimate crashing, change width to 90deg
Summary:
MCS tables from custom map profiles were not being sorted by range. This caused invalid geometry to be generated which caused turf to crash. This sorts the mcs ranges for both default profiles and custom profiles.

Also this changes the angle of the estimate to be 90 degrees instead of 30.

https://pxl.cl/1vw4t

Reviewed By: cjminer505

Differential Revision: D25640535 (https://github.com/facebookexternal/terragraph-apps/commit/1b794acdab9ea8054ca3b1eaf946255c5dcc2f0b)

fbshipit-source-id: 2bf2f633548e1d1d98ed8cb65b28ee798ae1694b