(scan service) Refactor scan execution status clean up logic.
Summary: It was observed that many scan tests were marked as FAILED even if we got partial scan results. The diff updates the execution status clean up logic. Now, we mark the test as FAILED only if none of the tokens were received, i.e mark the test as FINISHED if we receive at least one token.

Reviewed By: hmehmood0

Differential Revision: D27578030

fbshipit-source-id: 2cb7c9456b3d35898edb8a497bc8311233706434